ADd: navigation from CategoryCommonListFragment.kt to ProductListFragment.kt and  frpm ProductListFragment.kt to ProductFragment.kt

+++ app/src/main/java/com/ecoist/market/domain/repository/ProductRepository.kt
@@ -14,7 +14,7 @@ class ProductRepository(private val apiService: ApiService) {
         return ProductMapper.map(products)
     }
 
-    suspend fun getProductById(id: Int): Product {
+    suspend fun getProductById(id: Long): Product {
         val productResponse = apiService.getProductById(id)
         return ProductMapper.mapSingle(productResponse)
     }
